LayoutTests/media/media-startTime-expected.txt

Test the, so far unused, 'startTime' attribute.

EVENT(loadeddata)
EXPECTED (video.startTime == '0') OK
RUN(video.startTime = 10)
EXPECTED (video.startTime == '0') OK
END OF TEST
